Improve upon firefox home view controller + reloading section
