+++ Client/Frontend/Home/ASHeaderView.swift
@@ -6,7 +6,7 @@ import UIKit
 
 // MARK: - Section Header View
 public struct FirefoxHomeHeaderViewUX {
-    static let insets: CGFloat = UIDevice.current.userInterfaceIdiom == .pad ? FirefoxHomeUX.sectionInsetsForIpad + FirefoxHomeUX.minimumInsets : FirefoxHomeUX.minimumInsets
+    static let insets: CGFloat = UIDevice.current.userInterfaceIdiom == .pad ? FirefoxHomeViewModel.UX.sectionInsetsForIpad + FirefoxHomeViewModel.UX.minimumInsets : FirefoxHomeViewModel.UX.minimumInsets
     static let titleTopInset: CGFloat = 5
     static let sectionHeaderSize: CGFloat = 20
     static let maxTitleLabelTextSize: CGFloat = 55 // Style title3 - AX5
@@ -47,7 +47,7 @@ class ASHeaderView: UICollectionReusableView {
 
     var titleInsets: CGFloat {
         get {
-            return UIScreen.main.bounds.size.width == self.frame.size.width && UIDevice.current.userInterfaceIdiom == .pad ? FirefoxHomeHeaderViewUX.insets : FirefoxHomeUX.minimumInsets
+            return UIScreen.main.bounds.size.width == self.frame.size.width && UIDevice.current.userInterfaceIdiom == .pad ? FirefoxHomeHeaderViewUX.insets : FirefoxHomeViewModel.UX.minimumInsets
         }
     }
