+++ Client/Frontend/Home/HomePanelType.swift
@@ -1,0 +1,19 @@
+// This Source Code Form is subject to the terms of the Mozilla Public
+// License, v. 2.0. If a copy of the MPL was not distributed with this
+// file, You can obtain one at http://mozilla.org/MPL/2.0/
+
+import Foundation
+import Shared
+
+protocol HomePanel: NotificationThemeable {
+    var homePanelDelegate: HomePanelDelegate? { get set }
+}
+
+enum HomePanelType: Int {
+    case topSites = 0
+
+    var internalUrl: URL {
+        let aboutUrl: URL! = URL(string: "\(InternalURL.baseUrl)/\(AboutHomeHandler.path)")
+        return URL(string: "#panel=\(self.rawValue)", relativeTo: aboutUrl)!
+    }
+}
